refactor: change finder to have another state
This introduces another state to the finder that makes it start as a
`PreFinder` before it becomes an `ActiveFinder`. This also adds in a
makefile because it will save me maybe seconds of my life. I also added
in a formatting check for now just using cs launch until I can use the
built-in scala-cli check.